Separated code and adds login page view
