+++ src/ui/components/form/button/index.ts
@@ -1,0 +1,3 @@
+import ButtonComponent from './ButtonComponent';
+
+export default ButtonComponent;
