+++ src/ui/components/form/form-component/index.ts
@@ -1,0 +1,3 @@
+import FormComponent from './FormComponent';
+
+export default FormComponent;
