+++ src/ui/components/form/form-header/index.ts
@@ -1,0 +1,3 @@
+import FormHeader from './FormHeader';
+
+export default FormHeader;
