+++ src/ui/components/form/input-component/index.ts
@@ -1,0 +1,3 @@
+import InputComponent from './InputComponent';
+
+export default InputComponent;
