+++ src/ui/views/login/Login.tsx
@@ -1,14 +1,18 @@
 import React, { FunctionComponent } from 'react';
 
+import { FormProps } from '../../../interfaces/form/form.interface';
+import FormComponent from '../../components/form/form-component';
+import { LoginFormSetting } from './utilities';
+
 const Login: FunctionComponent = () => {
+  const formSettings: FormProps = LoginFormSetting;
+
   return (
-    <h1
-      style={{
-        textAlign: 'center',
-      }}
-    >
-      Log in here
-    </h1>
+    <FormComponent
+      SignUpFormSetting={formSettings}
+      buttonText="Login"
+      entryHeaderText="Login, blease!"
+    />
   );
 };
